+++ src/collective/elasticsearch/manager.py
@@ -45,6 +45,54 @@ class ElasticSearchManager:
         return value
 
     @property
+    def highlight(self):
+        """Is search highlighting enabled in the control panel."""
+        try:
+            value = api.portal.get_registry_record(
+                "highlight", interfaces.IElasticSettings, False
+            )
+        except KeyError:
+            value = False
+        return value
+
+    @property
+    def highlight_threshold(self):
+        """Search highlighting threshold."""
+        try:
+            value = api.portal.get_registry_record(
+                "highlight_threshold", interfaces.IElasticSettings, False
+            )
+        except KeyError:
+            value = 600
+        return value
+
+    @property
+    def highlight_pre_tags(self):
+        """Search highlighting pre tags."""
+        try:
+            value = api.portal.get_registry_record(
+                "highlight_pre_tags", interfaces.IElasticSettings, ""
+            )
+        except KeyError:
+            value = ""
+        if value is None:
+            value = ""
+        return value
+
+    @property
+    def highlight_post_tags(self):
+        """Search highlighting post tags."""
+        try:
+            value = api.portal.get_registry_record(
+                "highlight_post_tags", interfaces.IElasticSettings, ""
+            )
+        except KeyError:
+            value = ""
+        if value is None:
+            value = ""
+        return value
+
+    @property
     def catalog(self):
         return api.portal.get_tool("portal_catalog")
 
@@ -302,6 +350,12 @@ class ElasticSearchManager:
         if sort is not None:
             body["sort"] = sort
         warnings.simplefilter("ignore", ResourceWarning)
+        if self.highlight:
+            body["highlight"] = {
+                "fields": {"SearchableText": {"number_of_fragments": 5}},
+                "pre_tags": self.highlight_pre_tags.split("\n"),
+                "post_tags": self.highlight_post_tags.split("\n"),
+            }
         return self.connection.search(index=self.index_name, body=body, **query_params)
 
     def search(self, query: dict, factory=None, **query_params) -> LazyMap: